Moved components to its own location
